lotto_qr_app/requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
pyzbar>=0.1.9
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
